+++ src/OpenSymbolicRules.jl
@@ -1,7 +1,8 @@
 module OpenSymbolicRules
 
 using JSON
-import CommonSolve: solve
+import CommonSolve
+import CommonSolve: solve, init, solve!
 using PrecompileTools: @setup_workload, @compile_workload
 using SymbolicUtils
 using SymbolicUtils: @rule, Sym, Term